refactor(imports): remove file extensions from imports
