+++ src/components/atoms/index.ts
@@ -5,3 +5,4 @@ export { RowContainer } from './row-container';
 export { Separator } from './separator';
 export { SafeViewContainer } from './safe-view-container';
 export { AccountActionPressableContainer } from './account-action-pressable-container';
+export { SettingsMenuItem } from './settings-menu-item';
