+++ src/navigation/components/index.ts
@@ -1,2 +1,1 @@
 export { default as TabBar } from './tab-bar';
-export * from '../../screens/settings/components/settings-menu-item';
