--- src/screens/settings/components/index.ts
@@ -1,1 +1,0 @@
-export * from './settings-menu-item';
